requirements.txt
sentence-transformers==2.6.1
numpy>=2.0.0
fastapi>=0.110.0
uvicorn>=0.23.0
pdfminer.six>=20231228
python-docx>=1.1.0
PyMuPDF>=1.24.10
pdfplumber>=0.11.0
pillow>=10.3.0
pytesseract>=0.3.10
